Rename ReimbursementPerformer -> ReimbursementPerformer::Fifo
As per IRC chat w/ @cbrunsdon, @jhawthorn, @jdutil
